+++ packages/api/src/index.ts
@@ -6,4 +6,3 @@ export {
 } from "./version"
 export { constructExtensionSupportDir } from "./ui/server/path"
 export * from "./constants"
-export { TauriShellStdio } from "./runtime/stdio"
